fix: compilation errors, install Tauri deps, verified build
- Fixed u64/usize type mismatch in lib.rs
- Fixed typer_scaffold_lib → typer_lib in main.rs
- Fixed unused imports in adaptive.rs
- Fixed unused parameter and null check in main.ts
- Installed webkit2gtk and Tauri Linux deps
- Rust: 0 errors, 6 warnings (unused functions)
- TypeScript: 0 errors
- Binary: 12MB Tauri app builds successfully
- Can't run in container (no GTK display) but binary is valid

+++ src-tauri/src/lib.rs
@@ -129,7 +129,7 @@ fn record_keypress(
         confidence.insert(ch, lesson.key_confidence(ch, &stats));
     }
 
-    let new_position = position + 1;
+    let new_position = (position + 1) as usize;
 
     Ok(KeypressResult {
         correct,
